+++ test/utils/daos/test_userdao.py
@@ -1,11 +1,31 @@
 import os
+import re
 import pytest
 import sqlite3
 import unittest.mock as mock
-from src.utils.daos.userdao import UserDAO, UserData
+from src.utils.daos import UserDAO
 
 base_dir = os.path.dirname(__file__)
 test_db = os.path.join(base_dir, "test_data/test_db.sqlite")
+faked_rows = ((1, "johndoe", "author"), (2, "admin", "admin"), (3, "elmapelma", "moderator"))
+faked_names = (("id",), ("username", ), ("role", ))
+faked_expected = [
+  {
+    "id": 1,
+    "username": "johndoe",
+    "role": "author"
+  },
+  {
+    "id": 2,
+    "username": "admin",
+    "role": "admin"
+  },
+  {
+    "id": 3,
+    "username": "elmapelma",
+    "role": "moderator"
+  },
+]
 
 @pytest.fixture
 def sut_int():
@@ -24,17 +44,142 @@ def sut_int():
   conn.commit()
   conn.close()
 
-  with mock.patch("src.utils.daos.dao.os.environ.get") as db_path:
+  with mock.patch("src.utils.daos.basedao.os.environ.get") as db_path:
     db_path.return_value = test_db
-    sut = UserDAO()
+    sut = UserDAO("user")
     yield sut
     os.remove(test_db)
 
+
+@pytest.fixture
+def sut():
+  """SUT for unittest."""
+  sut = UserDAO("user")
+
+  return sut
+
+@pytest.mark.unit
+class TestUnitDAO:
+  @pytest.mark.parametrize("result, names, expected", [
+    ((2, "admin", "admin"),faked_names, {
+      "id": 2,
+      "username": "admin",
+      "role": "admin"
+    }),
+    (None, faked_names, None)
+  ])
+  @mock.patch("src.utils.daos.UserDAO._connect_get_cursor", autospec=True)
+  @mock.patch("src.utils.daos.UserDAO._disconnect", autospec=True)
+  def test_get_one(self, mockedDisconnect, mockedCGC, sut, result, names, expected):
+    """
+    Get one
+    """
+    mocked_cursor = mock.MagicMock()
+    mocked_cursor.description = names
+    mocked_cursor.fetchone.return_value = result
+
+    mockedCGC.return_value = mocked_cursor
+    mockedDisconnect.side_effect = None
+
+    result = sut.get_one(2)
+
+    assert result == expected
+
+  @mock.patch("src.utils.daos.UserDAO._connect_get_cursor", autospec=True)
+  def test_get_one_exception(self, mockedCGC, sut):
+    """
+    Get one - exception
+    """
+    mocked_cursor = mock.MagicMock()
+    mocked_cursor.fetchone.side_effect = Exception
+    mockedCGC.return_value = mocked_cursor
+
+    with pytest.raises(Exception):
+      sut.get_one(2)
+
+  @mock.patch("src.utils.daos.UserDAO._get_connection_and_cursor", autospec=True)
+  def test_update_exception(self, mockedCAC, sut):
+    """Test when an exception is raised due to wrong input, conn.close in finally block is called."""
+    connection = mock.MagicMock()
+    mockedCAC.return_value = connection, None
+
+    with pytest.raises(Exception):
+      sut.update(2, {})
+
+    connection.close.assert_called_once()
+
+  @mock.patch("src.utils.daos.UserDAO._get_connection_and_cursor", autospec=True)
+  def test_delete_exception(self, mockedCAC, sut):
+    """Test that when an exception is raised the _disconnect in finally block is called."""
+    connection = mock.MagicMock()
+    mockedCAC.return_value = connection, None
+  
+    with pytest.raises(Exception):
+      sut.delete(2)
+
+    connection.close.assert_called_once()
 
 
 @pytest.mark.integration
 class TestIntegrationUserDAO:
   """Integration tests."""
+  @pytest.mark.parametrize("id, expected",[
+    (1, "admin"),
+    (5, "johndoe")
+  ])
+  def test_get_one(self, sut_int, id, expected):
+    """Test get one user."""
+    data = sut_int.get_one(id)
+    
+    assert data["username"] == expected
+
+  def test_get_one_none(self, sut_int):
+    """Test when no match in database for one user."""
+    data = sut_int.get_one(2)
+
+    assert data is None
+
+  def test_create(self, sut_int):
+    """Test to create a user."""
+    input_data = {"username": "tony the tiger"}
+    data = sut_int.create(input_data)
+
+    assert "user_id" in data and input_data.items() <= data.items()
+
+  @pytest.mark.parametrize("input_data",[
+    ({"username": "admin"}),
+    ({"DROP TABLE user;": True})
+  ])
+  def test_create_fail(self, sut_int, input_data):
+    """Test to create a user with wrong input."""
+    with pytest.raises(Exception):
+      sut_int.create(input_data)
+
+  @pytest.mark.parametrize("id, input_data, expected",[
+    (5, {"username": "tony the tiger"}, True),
+    (2, {"username": "tony the tiger"}, False),
+  ])
+  def test_update(self, sut_int, id, input_data, expected):
+    """Test return value when updating a user."""
+    result = sut_int.update(id, input_data)
+
+    assert result == expected
+
+  @pytest.mark.parametrize("id, input_data, expected",[
+    (5, {"username": "tony the tiger"}, "tony the tiger"),
+  ])
+  def test_update_data(self, sut_int, id, input_data, expected):
+    """Test data when updating a user."""
+    result = sut_int.update(id, input_data)
+
+    conn = sqlite3.connect(test_db)
+    cur = conn.cursor()
+    cur.execute("SELECT username from user where id = ?", (id, ))
+    result = cur.fetchone()
+
+    conn.close()
+
+    assert result[0] == expected
 
   @pytest.mark.parametrize("user_name, expected_user_type", [
     ("admin", dict),
